Separated styling for pages. Curriculum and Project pages no longer conflitc.
