+++ src/components/CurriculumCard/CurriculumCard.css
