+++ src/components/CurriculumCard/CurriculumCard.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from "react-router-dom";
+import { Card } from "react-bootstrap";
 import './CurriculumCard.css';
 
 /*
