--- src/components/Curriculums.css
@@ -1,13 +1,0 @@
-.jumbotron {
-  width: 100%;
-  height: 300px;
-  margin: 0px;
-  background: #00B3FE;
-  color: white;
-}
-.jumbotron span {
-  vertical-align: middle;
-  line-height: 70px;
-  font-size: 70px;
-  overflow: scroll;
-}
