+++ src/components/Project.css
@@ -4,7 +4,6 @@
 	background-size: contain;
 	padding-top: 20%;
 	padding-bottom: 20%;
-
 }
 
 h2{
