+++ src/screens/Curriculum.js
@@ -17,14 +17,6 @@ class Curriculum extends Component {
   componentDidMount() {
     const { id } = this.props.match.params;
 
-    /*
-    Unused request for gitHub api option
-    axios.get(`https://api.github.com/repositories/${id}`, { headers: { Accept: 'application/vnd.github.mercy-preview+json' } })
-    .then(res => {
-      const curriculum = res.data;
-      this.setState({ curriculum });
-    }) */
-
     axios.get('/__mocks/curriculums.json')
       .then(res => {
           for (let o of res.data) {
